requirements_dev.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiohttp-retry==2.9.1
aiosignal==1.3.1
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.5.2
appdirs==1.4.4
asttokens==3.0.0
async-timeout==5.0.1
asyncssh==2.21.0
atpublic==5.0
attrs==25.3.0
backcall==0.2.0
backoff==2.2.1
backports.zoneinfo==0.2.1
billiard==4.2.1
black==22.8.0
blinker==1.8.2
boto3==1.37.38
botocore==1.37.38
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==2.2.1
colorama==0.4.6
comm==0.2.2
configobj==5.0.9
contourpy==1.1.1
cryptography==45.0.5
cycler==0.12.1
dacite==1.6.0
dagshub==0.5.10
dagshub-annotation-converter==0.1.11
databricks-cli==0.18.0
dataclasses-json==0.6.7
debugpy==1.8.15
decorator==5.2.1
dictdiffer==0.9.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docker==6.1.3
dpath==2.2.0
dulwich==0.22.4
dvc==3.42.0
dvc-data==3.8.0
dvc-http==2.32.0
dvc-objects==3.0.6
dvc-render==1.0.1
dvc-studio-client==0.20.0
dvc-task==0.40.2
entrypoints==0.4
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.16.1
flake8==5.0.4
Flask==2.3.3
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.44
gql==3.5.3
grandalf==0.8
graphql-core==3.2.6
greenlet==3.1.1
gto==1.6.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hydra-core==1.3.2
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.4.5
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.12.3
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.7
kombu==5.5.4
llvmlite==0.41.1
lxml==6.0.0
Mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.7.5
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mlflow==2.2.2
multidict==6.1.0
mypy==0.971
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.1
nltk==3.9.1
numba==0.58.1
numpy==1.24.4
oauthlib==3.3.1
omegaconf==2.3.0
orjson==3.10.15
packaging==23.2
pandas==2.0.3
parso==0.8.4
pathspec==0.12.1
pathvalidate==3.2.1
pickleshare==0.7.5
pillow==10.4.0
platformdirs==3.11.0
pluggy==1.5.0
prompt_toolkit==3.0.51
propcache==0.2.0
protobuf==4.25.8
psutil==7.0.0
pure_eval==0.2.3
py==1.11.0
pyarrow==11.0.0
pycodestyle==2.9.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydot==4.0.1
pyflakes==2.5.0
pygit2==1.13.3
Pygments==2.19.2
pygtrie==2.5.0
PyJWT==2.9.0
pyparsing==3.1.4
pytest==7.1.3
python-dateutil==2.9.0.post0
pytz==2022.7.1
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.0
querystring-parser==1.2.4
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.8
s3transfer==0.11.5
scikit-learn==1.3.2
scipy==1.10.1
scmrepo==2.0.4
seaborn==0.13.2
semver==3.0.4
shap==0.44.1
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
slicer==0.0.7
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
sqltrie==0.11.2
stack-data==0.6.3
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.3
tornado==6.4.2
tox==3.25.1
tqdm==4.67.1
traitlets==5.14.3
treelib==1.8.0
typer==0.16.0
typing-inspect==0.9.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
vine==5.1.0
virtualenv==20.31.2
voluptuous==0.14.2
waitress==2.1.2
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.6
xgboost==2.1.4
yarl==1.15.2
zc.lockfile==3.0.post1
zipp==3.20.2
